Add preliminary support for Aetherhub import
This commit setups the infrastructure needed to parse from aetherhub, in
particular it adds a parser, and a view that can be used to upload from
the parser. No effort is made at matching players from the tournament
report with the ones already in the database; we create new players for
everything.