Add support for calculating top 8 Qualification Points
This commit adds a new field to the model for results, storing results
after top8. It divides the results into "levels" in a single elmination
bracket, so we have one winner, one finalist, two semifinalists and 4
quarter finalists (for a total of 8).

There is still no support in the result importer for entering a top 8,
which will have to be done manually (top8s are usually handled outside
of the round management software, on paper by the head judge).

Another requirement is that we will have to store points for people who
are in the 9th-16th in a tournament. They do not play in the single
elimination portion of the tournament, but still get points based on
ranking. This means we will most likely have to store an additional
field "ranking" or something like that, containing results before top16.