Add a command to find duplicate players
This command can be used by the server admin to find people with similar
names. The idea here is that we can use it to find if there are
duplicate players who have been imported, and then to fix the
import code if required or to clean up the database from the admin
panel.