Add a description field to the event
Some tournament organizers might want to add their own description to
events, to directly guide players as to what to expect.

They also like to add some formatting to the list, like basic lists, so
we need to support basic HTML. However, we don't want to have XSS
attacks on our platform, that would be bad. So we escape the content of
the description, only allowing a very small set of tags. This is all
powered by the Python library Bleach [0], which seems to have a good
security track record. Their policy is that tags should be allowlisted
rather than blocklisted.

[0] https://pypi.org/project/bleach/