Display a user-friendly error on import failures
So far when we fail to import a tournament result, the only thing we get
is a "500 internal error" message, which is not very useful. This commit
changes that behavior to return 1) the correct HTTP error code (could be
improved) and 2) display a nice error message and return the user to the
edit form.