Add a view to list future events
In order to get some client side interactivity, we use a lightweight DOM
manipulation framework called Alpine. The reason to not go with
something heavierweight like React or Vue is that I did not want to have
a full on build step and complicated dependencies just to be able to
have some filters. The current solution seems like it would work well
enough, and should even be usable for some other pages (top8 for
example).