Add views for editing and deleting events
This commit add an important feature for TOs: the ability to either
delete or change their events. Up until now, this could only be
performed in the admin panel, but now it can also be done in the event
description page, which shows a link to a form to do this.