Correctly compute points for top16
Previously the code would always give extra points to top9-16 if the
event was large enough, independently of wether or not we played a top8
portion. This was due to a misunderstanding of the spec; the correct
interpretation is that the extra points are only awarded in case the
tournament played with a single elimination portion.

Note that this implementation strategy does not use denormalization; the
fact than an event played a top8 is extracted from the fact that it has
at least one result with a single elimination part filled in. This means
that in order to still have some performance, we need to use a lot of
aggregation. I am not sure wether de-normalizing this, by adding a
`has_top8` boolean field on events make the code simpler; it is another
source of issue that could arise when things get out of sync, but it
requires less aggregation to be useful.