Fix a bug with the future event view
The future event view was evaluating its queryset at the startup of the
app. This means the threshold date used for "today" was static, and was
the day the application was started.

This commit changes that, by putting the queryset in the get_queryset
function, which is called on every request, rather than be evaluated at
class instantiation time.