Disable caching of scores for now
We are observing some coherency issues with our scores cache, which I
believe is due to this. As we don't have a lot of results yet, this is
not causing some performance issues, and we will see how we can add a
better way around this in the future. I think one option might be to
have a more coherent cache, for example by running it in Redis or
similar.