Sort event results before displaying them
Currently event results were not really sorted, but rather shown as they
were created in the tournament management software. However, for events
with top8, it does not work that way; the standings can be different
from whatever happened in the top 8.

This commit introduces a sorting function for EventPlayerResults and
displays them sorted in events results.