Remove deprecated test
This test was intended to check that we could not create an event with
its round count set to a low value (less than 3 rounds). However, the
models don't have a round_count entry anymore, as those were removed
from the QPs algorithm.

What allowed the test to stay and pass was that we checked for
validation error on save, and the factory used did not pass a format to
events, triggering a validation error. However, once the factory was
fixed to create events with a format, the validation error would not
trigger anymore.

As the test is not testing anything useful anymore, delete it.