Display the top8 results in event details
When running events with top 8 results, we don't simply want to display
the ranking of a player. Instead, we care what they are doing in the
top8 portion of the tournament; where did they end up basically. Because
it is possible a player did finish first in the rounds of swiss but gets
kicked out of finals at the first top8 match.

This commit implements the changes required for this to work. It
basically sorts the results correctly, then extract everything with a
top8 result (which means we support top4 as well for smaller/team
events). For those event we display the single elimination result.