Remove some N+1 problematic queries
A bunch of views were running one query per object, which is far from
optimal. This was already fixed in the leaderboard view, as I was aware
that it was performance sensitive. However, this was not the case for
other views, which did not hit performance constraints.

For tracking down those queries, I used Pellet [0], which I might use
for exporting SQL query metrics to Prometheus later on for alerting on
performance regressions.

[0] https://github.com/har777/pellet