Refactor redirection after event upload
This refactor keeps the view more composable, by not returning a
response directly but instead getting the information from the base
classes. That way, behavior can be changed by mixins, should we wish
that.

We also do not manually construct an URL for the event anymore, but
instead defer to the absolute URL defined in the model and already used
by the admin panel for example.

Finally, this adds a test checking that we correctly redirect after a
succesful upload.