Cleaning up PARSERS_LIST
First of all, rename parsers into importers: parsers are function taking
a string and returning structured data. Those modules do a bit more than
that, as they are responsible for actually creating items in the DB. The
files named parsers will still remain, as they are actually doing
parsing.

In addition, restructure a bit the lists of parsers so that it is harder
to have information that is out of order by tying them together (lists
of structs vs structs of list).

Finally, minor cleanups around properties.