Limit maximum age of events when uploading
We don't want TOs to be able to enter very old results (at least not
without asking) so that the ranking doesn't suddenly change at the end
of the year.

This commit also introduces a model manager for Events, allowing us to
expose custom query set, here the set of events for which we could
upload results.