Make top8 mandatory above 16 players
Appendix E of the Magic Tournament Rules say that playing a full top8 is
recommended above 16 players, and that top4 is fine below that. The
league also uses that rule, so make the quarterfinalist field required
depending on the number of players. The form validation will handle the
rest for us.