Sort results by points before storing
Currently results are assumed to be entered in ranking order which might
or might not be the case when uploading results generated from something
else, like using Round Robin scheduling.

Note that Python's sort method is guaranteed to be stable, meaning that
for points tie, ranking will be mased on input order.