Start working on generating invoices
I choose to put invoice management in a separate app as it felt quite
different from the general championship-oriented functionalities;
basically all the invoice management could have been done either via an
offline tool, so it made sense to me to separate it into a different
app.

I define a new model used to represent the informations required for
creating an invoice PDF. The idea is that this way, we don't need to
keep track of the metadata somewhere else, or to provide an in-between
somewhere else.

I also generate the view needed to create a PDF from a given invoice,
using latex as a backend for rendering the PDF.