Disable results validation on some events
This feature is requested for some events (i.e. Swiss Magic Masters).
While the feature is generally useful, it can get in the way, for
example if an event has players with byes, it could end up with more
points in total than plausible, but it is still something we should
allow.

Note that this does not try to disable specific checks: we assume that
those events are going to be very few and in-between, so we disable
everything and use caution / rely on manual inspection of results.