Store record (W/L/D) in addition to match points.
This is step two of the plan to convert our internal representation from
match points to win loss draw. Now we will start writing to the database
in both formats, points and record to check that this works correctly.
No computation uses the new format anymore, it is only stored.

Next steps:

- Convert existing database to the new format.
- Delete the "points" column and convert it into a property instead.