Migrate data from points to record
This migration adds transforms the existing data from points to match
record. It runs this in a Django migration, meaning the data migration will
automatically run as part of the CI/CD process. It also includes a way
to rollback this migration if we ever need to. I tested both by running
it on my local copy of the database.

I copied some of the code rather than import it because I don't
want the result of this migration to change in the future if someone
changes the estimation function.