Simplify the template selection logic
Instead of redifining the entierty of the HTTP GET handler, which is not
really needed, we can simply override the part of the code that chooses
which template to use.

Also, instead of checking manually which template exists and how to
fallback, we can provide Django with several choices and it will use the
first one which exists.