Wrap all the season-specific scores in a class
As discussed with Jari, we don't want to have many different
season-specific methods for computing qps, and byes, etc.

Instead, let's define new per-season types that hold all of that
season's scoring metrics, and they can be aggregated together by
overriding the + operator.

This commit is not completely season generic yet, some type signatures
have to be changed and the computation function has to be chosen based
on the season chosen, but this is for a future commit.