Add a new PayeeAddress model
This model will be used to store banking contact for invoices. It will
allow us to create different invoices which should be paid for different
people. We need this at the end of the 2023 season as the finances of
the SUL get handed to a new person / account.

This commit also contains a migration which creates a new PayeeAddress
with the address that was hardcoded in our template and attaches it to
existing invoices, so that existing invoices don't change from a TO
point of view.