Exclude some events from invoices
Some of our events are a bit special and should not appear as bill
items. For example, the SUL finals are technically a premier event
organized by Leonin League, but we are not collecting SUL fees for it.

This commit adds a boolean field allowing a superuser to go and mark an
event as "do not bill".