Implement sending emails from prod
This commit reads the API key which is injected in the container's
environment variables (already implemented on the server), and sets it
up for integration with sendgrid.

I recently set up the sendgrid integration for unityleague.ch domain,
but I am still waiting for it to propagate, before I can really test
things out.