Add the ability to freeze invoice PDF
Currently, maintaining the invoices PDF is a bit dangerous; it could be
quite easy to introduce a change in the PDF generation code that could
change the output of the generated invoice, which is bad for our TOs,
who would like their invoices to not change ideally.

To make our lives easier, we introduce a FileField attached to invoices
which can store the generated PDF. This is *not* a cache: it is not here
for performance reason, but for avoiding unexpected changes to the PDF.

The freeze / unfreeze aspect can be controlled via a new admin action.