Allow blank frozen files in invoices
Currently, it is not possible to manually create an invoice in the
Django admin, as the file is not allowed to be left blank in forms, even
though it is intended to be blank.

And while we are at it, remove it from the admin page, as it should not
be editable by a user, only through the special actions.