Extract API to a separate app
Currently the API is included in the "championship" app, which roughly
includes everything running on unityleague.ch, so creating the new
events, importing results, and so on.

However, I would like in the future to implement some API for the
invoicing, or possibly for future features such as decklists, and I
think it does make sense to have it in a separate app that imports
models from championship and invoicing and so on, as needed.

This commit does not change any code, just move things around.