Convert PerSeasonView into a mixin
Currently, the per season "add on" can only be used with TemplateViews,
which breaks using it with some of Django's class based views (like
Detail view).

This commit switches it to be a mixin, which does not handle any of the
actual rendering, but simply adds the season information into the mix.