Work around flaky test
The test that we are modifying here was flakey as it would stop working
if the PlayerFactory picked a short first name, which in turn would get
mangled when running through clean_name.

The workaround for now is to force-set the name to use in this test. I
also left a TODO to properly address this in the clean_name function, as
I cannot really do it for now.