Show Premier events before regional on index page
Currently on the website we have two events by Azcanta tournaments on
the same day as they are organizing a rebound modern after their
premier. Unfortunately, the rebound is listed first on the index page,
while it would make more sense to show the premier first. With the
current code, regionals will always be shown first.

This commit changes the sort logic to take into account the category
when sorting.