Guarantee unicity of card name
This is a quick workaround the issue we had where card names returned by
the factory are not unique. As a change we instead take a sequence of
card names, and loop around once we used too many cards (which should
not happen in a single test).

Original idea by Jari Rentsch.