Invoice admine: Store total amount in cache
The total amount is a useful property to show in the overall view, as it
allows one to check from the bank statement which invoice it is.

However, loading this total amount is also very slow, as it requires
checking every event's player count. So we will store it in a cache, so
that when using the website to enter who paid, it is reasonnably snappy.