Rename EventPlayerResult to just Result
EventPlayerResult as a name is needlessly verbose, requires more typing
and does not really give us more meaning than just "Result". The only
upside is that we somehow reflect in the name that Event and Player are
linked through the EPR, but once you learn that Result is the link
between the two, it should be enough.