Expose the revenue plot as a web view
This commit adds a simple visualisation of the revenue plot I sometimes
shared in our whatsapp chat directly on the web app. This will help me
spend less time doing this plot, but will also allow our budget /
accounting folks to look at it themselves and in realtime.

The code for the financial report should allow us to extend it with
different type of plot later, if required.

This page is protected by a login as I don't think we should be exposing
this to everyone, even though we are quite transparent about our
finance.