Transfer Docker image through artifacts
The previous implementation used Kaniko to login into a registry and
push the image there, while the rollout job would pull the image. This
had some downsides:

- It requires the CI server to have a login to the registry.
- It requires a registry which holds our code (not a downside as we are
  open source, but still a dependency).
- It prevents us from running the build step on other branches than
  master.
- It consumes bandwidth to another machine.

The new implementation instead uses the Gitlab artifact store to hold an
archive of the Docker image and push it to the rollout job.