Add a field describing the purpose of the discount
In order to make more accurate statistics and track why we give
discounts to people, add a new field on the Invoice model. This flag is
set as an Enum rather than as a boolean, even though there is only two
values for now, as we may find that we want to have more options later.