Take discount into account when plotting revenue
This commit takes into account discounts we gave to tournament
organizers when plotting the revenue of the SUL over the whole season.
This gives us a more accurate comparison of seasons, as, for instance,
the cheaper events of 2024 led to less discounts given out to TOs.

As discounts are associated with invoices rather than events, we first
needed to find a date where this discount was to be taken into account.
While this will never be 100% accurate, Jari suggested that we apply the
discount on the first Premier of the invoice, as this is usually the
event that TOs wanted discounted. When they don't have a premier, we
apply it on the first Regional instead.

One downside of this approach is that currently the plot for the revenue
can sometimes go down slightly, when the first event of an invoice
brings less money than the invoice's discount. This is only cosmetic
though, and eventually the data lines will converge to the reality.

Another issue with the current approach is that we end up doing one SQL
query per invoice with a discount. While this might theoretically be an
issue as we add more invoices, I think the number of invoices should be
low enough that the added complexity of a smarter approach is not worth
it.