Start working on a blogging system
This commit marks the first step towards supporting basic articles on
the Swiss Unity League website, as requested from a contributor.

Articles will be stored not in the application source code like info
pages, but in an article table. The reason for this is that we expect
articles to be more often changed, including by non-programmers.