Implement card modal for decklists
This commit adds a modal so that when a user clicks on a card they are
displayed with the art of the card as well as a link to Scryfall. For
now this is only implemented in the decklists viewer, but the frontend
code is reusable for the articles in the future.