Add the option to search for an exact card name
Doing an exact card name lookup is much faster than searching for case
insensitive, since it can be done with an index rather than a full table
scan (from ~7ms to < 1ms).

Since this cannot be done in a decklist context (because some players
will mess it up), control it through a flag to enable it only when done
in an article.