Implement decklist support in articles
It is now possible to reference a decklist stored in the SUL system in an
article, which will then get displayed in a nice box in the article,
with link to the cards and so on.

Note that while this might not be the fastest way to render a decklist
(lots of queries to the oracle DB), we can always cache the article
content, which we will do if we realize it is an issue. Some early
testing on my machine showed rendering time of 300 ms for one article,
which can get shut down to ~40 ms. I will not implement it for now, as I
would like real numbers from a real article on the real server before
deciding to invest time on this (in particular, cache invalidation is
not easy).