Define the 2025 season dates
This commit adds the 2025 season information to season.py. In order not
to break any existing code, the season is added as "invisible", meaning
the functions to find it by date or by slug do exist, but it is not
exposed in any public page.

This commit also adds the relevant factory for events in 2025.