Separate cleanly between public and private APIs
The "private" API here means the API used by our only dynamic view, the
upcoming event view. The reason for separating the two is to highlight
the fact that the code in api/ is used by other people to connect to our
website (or at least in theory could be used to do that), while the code
in championship/ can be freely modified by us to add or remove fields.

Conceptually, it also makes more sense that code only used by one
championship view lives in championship, while public API code lives under
api/. Finally, it also removes the private API views from the API router
exposed at unityleague.ch/api.

This commit as a byproduct also deletes an autocomplete view that is not
used anywhere, as far as I can tell.