Add a sharing link to collections
This commit introduces a "sharing link" to the decklist collections,
allowing the organizer to share the list of players without having to
look at them by themselves, for instance to share it with a caster crew
or with a judge, without them needing to have a login on unityleague. Of
course, we cannot prevent the organizer themselves from looking, but at
least it requires an additional click.

This is implemented with a "sharing key", 128 bit passwords that can be
added to the GET request for the Collection as a way to "unlock" the
access to the links. Doing it like that instead of through a primary key
keeps the URL stable and avoids introducing another key that would need
to be indexed for instance.