Add a query parameter for sorting mode
This commit introduces a query parameter for sorting decklists, either
by creatures (by default, most interesting to players), or by mana value
for deck checks (most interesting to judges). In the future we could
imagine having some kind of UI to switch between the two.

When in the "judge view", decklists links are setup to sort by mana
value.