Speed up leaderboard filtering with Alpine
This commit replaces the hand-written search functionality with one
based on rendering the leaderboard through an Alpine.js template,
similar to how we do it for the event filter.

A difference between this implementation and the event one is that here
we do not have a new API handler, but instead embed the required data in
the page using Django's `json_script` approach, which I was not aware of
when we implemented events.

The speedup is quite impressive, at the cost of a slight increase in
required code. We do not add new dependencies though.