Correctly display results for SUL Other
While SUL Other events were not supposed to have any Results attached to
them, we ultimately used them to represent the SUL Invitational, so we
need to be able to attach standings to those.

The issue is that get_results_with_qps filters out events with SUL
Other before passing them down to the Score methods, so that the methods
don't have to handle this category themselves. Workaround is to simply
attach None as score. This is not very elegant, but is fairly localized
to a single view.