Introduce a simple case for site filtering
This commit introduce a filter on the /organizer page to only display
organizers on the current site. This is low-risk enough that it can be
merged for now and displays how to use such a field.

Also introduce a filter for Events on site, although we will not need it
for now.