Display OTHER events in event history
We are now using OTHER events to represent events that do not award
points, but some of them can still have results attached to them (e.g.
the 2024 invitational). Therefore, display it in the player's event
history, with "N/A" as score.

Also introduce some result query set for making querying for a season
more readable.