Cleanup a bit the information view
The current code made the views depend on the URL config, by splitting
up the path which is bad practice in Django (what happens if you move
your page?). This commit replaces this compicated logic by inheritance.

Also cache the result of template_exists for good measure as it will be
cheap.