multisite: Create a model for site-specific infos
This commit creates a new model in the database linked to the site and
used to store current information about the website. It is kind of
similar to the User/UserProfile link that is often used in Django to tie
together a django.contrib.auth.models.User and additional information.