Adds GPS coordinates to Address
This commit adds GPS coordinates to Addresses and overrides the save()
method so that when an address is edited, it automatically gets an
updated GPS coordinate (we don't expect this to happen often).

Storing the GPS coordinates in the DB is required for efficient lookup
of events near a user, which is a feature that will gets added later.

This commit also contains a migration intended to geocode all existing
addresses in the system, as well as some test utilities to avoid
geocoding while we are in a unit test.